Make sure that MemoryAllocator uses wide-enough unsigned types for representing the address space being allocated
The types should be able to represent any address in the address space of the device, whose memory is being allocated by means of the MemoryAllocator. This device may be different from the host running Glow. E.g. Glow may run on a host with a 32-bit address space, but use a  MemoryAllocator for a GPU or an accelerator with a 64-bit address space.

Therefore, to be on the safe side use `uint64_t` instead of `size_t` in MemoryAllocator APIs.

Update the user's of the API to use `uint64_t` instead of `size_t`.

+++ lib/Backends/CPU/AllocationsInfo.cpp
@@ -101,7 +101,7 @@ void AllocationsInfo::allocateActivations(const IRFunction *F) {
   MemoryAllocator activationsAllocator(0);
 
   // Maps activations and views to some offset within the heap.
-  llvm::DenseMap<const Value *, size_t> activationAddr;
+  llvm::DenseMap<const Value *, uint64_t> activationAddr;
 
   // Assign device-space addresses to the activations.
   for (const auto &I : F->getInstrs()) {
